Make the following changes:
,

Find this element:

  <CardContent>

with the text content:

    Your Estimated Total Daily Macros:Calories: 1961 kcalProtein: 140.0 gCarbs: 229.0 gFat: 54.0 g

Relevant files:

    - https://6000-firebase-studio-1747278181438.cluster-f4iwdviaqvc2ct6pgytzw4xqy4.cloudworkstations.dev/_next/static/chunks/src_25bd91a2._.js
    - https://6000-firebase-studio-1747278181438.cluster-f4iwdviaqvc2ct6pgytzw4xqy4.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js
    - https://6000-firebase-studio-1747278181438.cluster-f4iwdviaqvc2ct6pgytzw4xqy4.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

I want these numbers to be received from Daily macro breakdown, if that is empty, get the numbers from smart aclorie planner. Also, put a comment here so the users know where to adjust these numbers.